Modify `new` function of Permutation
The new function gets thenumber of full rounds as
the ammount of full rounds that it will do before
and after the partial rounds.

This allows us to get rid from the result and the error.

Also modified the `Default` impl values of the full_rounds
parameter to 4.

Finally, removed the divisions by 2 on the `result` function
implementation.