Change names according to Poseidon paper
`R_f` is now the half of the total full rounds applied
on the Permutation resolve process.

`R_p` is ho the paper calls to `Partial rounds`.

Modified the `result()` method in order to iterate
through the half of the full rounds before and after
the partial ones.

This closes #2