Make quintic_sbox pass tests
Previously the code was executing `s = s*s*s*s*s;`
where `s` was previously derefed and is mutable.

My thoughts are that the muls were accumulating instead
of mul the same value 5 times.

Switch to `pow` method of BLS solved the err.

Currently quintic_sbox is being computed correctly in both
Strategy implementations and the tests are passing.