Fix hades_preimage test bug
The test was checking that the Variables stored inside the
`perm` object where the same as the output obtained using
the `Scalar` strategy.

The problem is that these values were not the correct ones
since the `perm()` fn actually changes the &mut input instead
of storing any results.

So we should copy the values inside of the perm after it was
executed instead of doing it before. On this way, we get the
input permuted which is what we want.

Fixes a bug that was blicking #43